Update `gomplate` code
